+++ src/presentation.extension/services/iExtensionServices.ts
@@ -1,21 +1,6 @@
 import { DependencyCache } from 'domain/packages';
 import {
-  OnActiveTextEditorChange,
-  OnClearCache,
-  OnErrorClick,
-  OnFileLinkClick,
-  OnPackageDependenciesChanged,
-  OnPreSaveChanges,
-  OnProviderEditorActivated,
-  OnProviderTextDocumentChange,
-  OnProviderTextDocumentClose,
-  OnSaveChanges,
-  OnTextDocumentChange,
-  OnTextDocumentClose,
-  OnTextDocumentSave,
-  OnTogglePrereleases,
-  OnToggleReleases,
-  OnUpdateDependencyClick,
+  IEventServices,
   SuggestionCodeLensProvider,
   SuggestionsOptions,
   VersionLensExtension,
@@ -23,8 +8,7 @@ import {
 } from 'presentation.extension';
 import { OutputChannel } from 'vscode';
 
-export interface IExtensionServices {
-
+export interface IExtensionServices extends IEventServices {
   suggestionOptions: SuggestionsOptions,
 
   extension: VersionLensExtension;
@@ -36,39 +20,4 @@ export interface IExtensionServices {
   versionLensProviders: Array<SuggestionCodeLensProvider>;
 
   editorDependencyCache: DependencyCache;
-
-  // vscode events
-  onActiveTextEditorChange: OnActiveTextEditorChange;
-
-  onTextDocumentChange: OnTextDocumentChange;
-
-  onTextDocumentClose: OnTextDocumentClose;
-
-  onTextDocumentSave: OnTextDocumentSave;
-
-  // version lens events
-  onProviderEditorActivated: OnProviderEditorActivated;
-
-  onProviderTextDocumentChange: OnProviderTextDocumentChange;
-
-  onProviderTextDocumentClose: OnProviderTextDocumentClose;
-
-  onPackageDependenciesChanged: OnPackageDependenciesChanged;
-
-  onClearCache: OnClearCache;
-
-  onPreSaveChanges: OnPreSaveChanges;
-
-  onSaveChanges: OnSaveChanges;
-
-  onFileLinkClick: OnFileLinkClick;
-
-  onUpdateDependencyClick: OnUpdateDependencyClick;
-
-  onToggleReleases: OnToggleReleases;
-
-  onTogglePrereleases: OnTogglePrereleases;
-
-  onErrorClick: OnErrorClick;
-
 }